+++ src/js/util/class.js
@@ -1,5 +1,6 @@
+import {isIE} from './env';
 import {attr} from './attr';
-import {includes, isString, isUndefined, toNodes} from './lang';
+import {includes, isArray, isString, toNodes} from './lang';
 
 export function addClass(element, ...args) {
     apply(element, args, 'add');
@@ -34,11 +35,11 @@ export function toggleClass(element, ...args) {
 
     args = args.filter(Boolean);
 
-    toNodes(element).forEach(({classList}) => {
+    toNodes(element).forEach(element => {
         for (let i = 0; i < args.length; i++) {
-            supports.Force
-                ? classList.toggle(...[args[i]].concat(force))
-                : (classList[(!isUndefined(force) ? force : !classList.contains(args[i])) ? 'add' : 'remove'](args[i]));
+            !isIE
+                ? element.classList.toggle(...[args[i]].concat(force))
+                : classList.toggle(element, args[i], force);
         }
     });
 
@@ -47,10 +48,10 @@ export function toggleClass(element, ...args) {
 function apply(element, args, fn) {
     args = getArgs(args).filter(Boolean);
 
-    args.length && toNodes(element).forEach(({classList}) => {
-        supports.Multiple
-            ? classList[fn](...args)
-            : args.forEach(cls => classList[fn](cls));
+    args.length && toNodes(element).forEach(element => {
+        !isIE
+            ? element.classList[fn](...args)
+            : classList[fn](element, args);
     });
 }
 
@@ -60,18 +61,28 @@ function getArgs(args) {
         , []);
 }
 
-const supports = {};
+// IE 11 shim (SVG elements do not support `classList` in IE 11)
+const classList = {
 
-// IE 11
-(function () {
+    add(element, classNames) {
+        classNames.forEach(name => !this.contains(element, name) && attr(element, 'class', `${attr(element, 'class')} ${name}`));
+    },
 
-    let list = document.createElement('_').classList;
-    if (list) {
-        list.add('a', 'b');
-        list.toggle('c', false);
-        supports.Multiple = list.contains('b');
-        supports.Force = !list.contains('c');
+    remove(element, classNames) {
+        classNames.forEach(name => attr(element, 'class', (attr(element, 'class') || '').replace(new RegExp(name, 'g'), ' ').trim()));
+    },
+
+    toggle(element, className, force) {
+        this[
+            (!isArray(force)
+                ? force
+                : !this.contains(element, className)
+            ) ? 'add' : 'remove'
+        ](element, [className]);
+    },
+
+    contains(element, name) {
+        return (attr(element, 'class') || '').match(new RegExp(name));
     }
-    list = null;
 
-})();
+};