+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 
 - Fix filter component initial active state
 - Fix translating `x` and `y` properties simultaneously in Parallax component
+- Class utils can handle SVGs in IE 11 now
 
 ## 3.0.3 (January 29, 2019)